Added a hack to allow running with "-enable-kvm".
The hack is to remove the NX bit from P4D in the page table. This allows
executing the windows loader code in a memory space which is seemingly
user-space. Since IA32_EFER.NXE == 1, having P4D with the NX bit will
result in a page fault, when we start fetching instructions for Windows
loader.